Adding Modal button
Created a new folder "Modal" in src/components
Added two new files to folder "Modal" named Settings.js and Modal.js
Edited index.js to add "import Settings from './components/Modal/Settings';" and to add <Settings /> to the render.
Edited the index.html to add "<div id="portal"></div>"

I have added a "Settings button" to the top left of all pages, that will open a modal on all pages. This adds a black overlay of 30% to the web page, making the modal appear at the forefront. The modal contains a button to close it.