Addition of Leave button and Volume Slider to settings menu
Have added a leave button to the settings menu (settingsModal.js) that will take the user back to the home.js page. The URL must be changed for each user as it currently redirects to my local host (3002).

The volume slider has also been added to SettingsModal.js. However, the fucntion that allows it to move is not currently in the correct place, but i am just encountering errors wherever i place it, so i have commented it out.

Have created an additional files in the SettingsMenu folder, being index.css, which contains the css for the volume slider.

Have edited the width, height and text alignment for the settings overlay.